Modularize main/index.ts into focused modules
